Animation splash for the start of the app. The app icon has been change to our Logo. To have the splash in the starting of the app, we needed to create a new activity and class for the previous Mainactivity now call MainPage. And the Splash animation is in the MainActivity class and layout. Related issue: #29  CoAuthor: #Quinstedt
