+++ Android/islandRush/app/src/main/res/anim/car_start_animation.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+
+        <translate android:duration ="5000"
+            android:fromXDelta="-20%p" android:toXDelta="120%p"
+            android:fromYDelta="50%p" android:toYDelta="50%p">
+
+        </translate>
+
+</set>
