+++ Android/islandRush/app/src/main/res/anim/start_animation.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+<rotate android:duration ="5000" android:fromDegrees="0.0"
+    android:pivotX="50%" android:pivotY="50%" android:toDegrees="360">
+
+</rotate>
+</set>
