+++ Android/islandRush/app/src/main/res/drawable/animation_background.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape>
+            <gradient
+                android:angle="255"
+                android:startColor="#0DDFCD"
+                android:endColor="#091B8E"
+                />
+        </shape>
+    </item>
+</selector>
