+++ Android/islandRush/app/src/main/res/layout-land/activity_main.xml
@@ -1,51 +1,92 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mainLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:background="@drawable/dynamic_list">
+    tools:context=".MainActivity">
 
-    <EditText
-        android:id="@+id/playerName"
-        android:layout_width="326dp"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_marginStart="50sp"
-        android:layout_marginEnd="50sp"
-        android:autofillHints="Player Name"
-        android:hint="Enter Name"
-        android:inputType="text"
-        android:textAlignment="center"
-        android:minHeight="48sp"
-        android:textSize="20sp"
-        android:textColor="@color/white"
-        android:textColorHighlight="#2CDDCF"
-        android:textColorHint="@color/white"
-        android:textColorLink="#08BCD8" />
+    <ImageView
+        android:id="@+id/palmBackground"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/animation_background"
+        android:contentDescription="@string/palm_image"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.471"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.282"
+        app:srcCompat="@drawable/palm" />
 
+    <ImageView
+        android:id="@+id/carImage"
+        android:layout_width="319dp"
+        android:layout_height="126dp"
+        android:background="@android:color/transparent"
+        android:contentDescription="@string/car_image"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.427"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.196"
+        app:srcCompat="@drawable/bil" />
 
-    <Button
-        android:id="@+id/button_enterRace"
+    <androidx.constraintlayout.widget.Guideline
+        android:id="@+id/guideline2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@id/playerName"
-        android:layout_centerInParent="true"
-        android:text="  Enter Race  "
-        android:textSize="20sp" />
+        android:orientation="horizontal"
+        app:layout_constraintGuide_begin="-7dp" />
 
-    <Button
-        android:id="@+id/button_Leaderboard"
+    <androidx.constraintlayout.widget.Barrier
+        android:id="@+id/barrier2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="LeaderBoard "
-        android:layout_below="@+id/button_enterRace"
-        android:layout_centerInParent="true"
-        android:textSize="20sp"
-        />
+        app:barrierDirection="top"
+        tools:layout_editor_absoluteY="1143dp" />
+
+    <TextView
+        android:id="@+id/animationTitle"
+        android:layout_width="wrap_content"
+        android:layout_height="70dp"
+        android:background="#070071D4"
+        android:text="@string/islandrush"
+        android:textAllCaps="true"
+        android:textAppearance="@style/TextAppearance.AppCompat.Medium"
+        android:textColor="#FF9800"
+        android:textSize="40sp"
+        android:textStyle="bold"
+        android:visibility="visible"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.049" />
+
+    <androidx.constraintlayout.helper.widget.Flow
+        android:id="@+id/flow"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+
+    <androidx.constraintlayout.widget.Guideline
+        android:id="@+id/guideline3"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        app:layout_constraintGuide_begin="0dp" />
+
+    <androidx.constraintlayout.widget.Guideline
+        android:id="@+id/guideline4"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        app:layout_constraintGuide_begin="805dp" />
 
 
 
-</RelativeLayout>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
