+++ Android/islandRush/app/src/main/res/layout-land/activity_main_page.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainPage"
+    android:id="@+id/mainPage"
+    android:background="@drawable/dynamic_list">
+
+    <EditText
+        android:id="@+id/playerName"
+        android:layout_width="326dp"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:layout_marginStart="50sp"
+        android:layout_marginEnd="50sp"
+        android:autofillHints="Player Name"
+        android:hint="Enter Name"
+        android:inputType="text"
+        android:textAlignment="center"
+        android:minHeight="48sp"
+        android:textSize="20sp"
+        android:textColor="@color/white"
+        android:textColorHighlight="#2CDDCF"
+        android:textColorHint="@color/white"
+        android:textColorLink="#08BCD8" />
+
+
+    <Button
+        android:id="@+id/button_enterRace"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/playerName"
+        android:layout_centerInParent="true"
+        android:text="  Enter Race  "
+        android:textSize="20sp" />
+
+    <Button
+        android:id="@+id/button_Leaderboard"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="LeaderBoard "
+        android:layout_below="@+id/button_enterRace"
+        android:layout_centerInParent="true"
+        android:textSize="20sp"
+        />
+</RelativeLayout>
